requirements.txt
torch==1.12.1
torchvision==0.13.1
numpy==1.21.6
scikit-learn==1.0.2
pandas==1.3.5
pyyaml==5.4.1
pyhopper